Allow opting out of settings deny_unknown_fields for foundations internal structs (#85)
Following up on #49 

It's been realized that there is no existing way to opt out of
`deny_unknown_fields` for the settings structs defined within
foundations itself. This adds a way to do that - by settings
`default-features = false` in `Cargo.toml` and not passing the new
`settings_deny_unknown_fields_by_default` feature.

This keeps it enabled by default, but gives users of foundations a way
to disable this feature. When the feature is not passed, since it only
controls the default structs that don't specify, it is still possible to
opt specific settings structs into this check.